requirements.txt
torch==1.13.1
opencv-python==4.5.5.62
open3d
loguru
torchmetrics
torchtyping
nerfacc==0.3.5
gin-config==0.5.0
termcolor==2.4.0
Pillow==9.5.0
